+++ 02_munge/src/munge_usgs.py
@@ -3,16 +3,29 @@ import numpy as np
 import pandas as pd
 import re
 import boto3
+import sys
 
-# get dictionary of all possible USGS site parameters
+def prep_write_location(write_location):
+    if write_location=='S3':
+        cont = input("You are about to write to S3, and you may overwrite existing data. Are you sure you want to do this? (yes, no)")
+        if cont=="no":
+            sys.exit("Aborting data fetch.")
+        # start S3 session so that we can upload data
+        session = boto3.Session(profile_name='dev')
+        s3_client = session.client('s3')
+    else:
+        s3_client = None
+    return s3_client
+
 def process_params_to_csv(raw_params_txt, params_outfile_csv, s3_client):
     '''process raw parameter text file into a csv file'''
     print('processing parameter file and saving locally')
     params_df = pd.read_csv(raw_params_txt, comment='#', sep='\t', lineterminator='\n')
     params_df.drop(index=0, inplace=True)
     params_df.to_csv(params_outfile_csv)
-    print('uploading to s3')
-    s3_client.upload_file(params_outfile_csv, 'drb-estuary-salinity', '02_munge/out/'+os.path.basename(params_outfile_csv))
+    if s3_client:
+        print('uploading to s3')
+        s3_client.upload_file(params_outfile_csv, 'drb-estuary-salinity', '02_munge/out/'+os.path.basename(params_outfile_csv))
 
 def process_data_to_csv(raw_datafile, flags_to_drop, agg_level, prop_obs_required, s3_client):
     '''
@@ -56,14 +69,15 @@ def process_data_to_csv(raw_datafile, flags_to_drop, agg_level, prop_obs_require
     # save pre-processed data
     data_outfile_csv = os.path.join('.', '02_munge', 'out', os.path.splitext(os.path.basename(raw_datafile))[0]+'.csv')
     df.to_csv(data_outfile_csv, index=True)
-
-    print('uploading to s3')
-    s3_client.upload_file(data_outfile_csv, 'drb-estuary-salinity', '02_munge/out/'+os.path.basename(data_outfile_csv))
+    
+    if s3_client:
+        print('uploading to s3')
+        s3_client.upload_file(data_outfile_csv, 'drb-estuary-salinity', '02_munge/out/'+os.path.basename(data_outfile_csv))
 
 def main():
-    # start S3 session so that we can upload data
-    session = boto3.Session(profile_name='dev')
-    s3_client = session.client('s3')
+    # choose where you want to write your data outputs: local or S3
+    write_location = 'S3'
+    s3_client = prep_write_location(write_location)
 
     # process raw parameter data into csv
     raw_params_txt = os.path.join('.', '01_fetch', 'out', 'usgs_nwis_params.txt')
